refactor: getStoreDetail 리팩토링
