+++ src/api/store/index.ts
@@ -1,4 +1,5 @@
 export * from './types';
 export * from './useGetStore';
-export * from './useGetStoreDetail';
+export * from './getStoreDetail';
+export * from './getStoreMinimum';
 export * from './useGetStoreMetadata';
