+++ src/components/store/StoreDetailWindow.tsx
@@ -31,7 +31,7 @@ const StoreDetailWindow = (props: StoreDetailWindowProps) => {
         fontWeight='bold'
         sx={{ borderBottom: '1px solid', borderColor: 'divider' }}
       >
-        {storeDetailData?.storeData.store_name}
+        {storeDetailData?.storeData?.store_name}
       </Typography>
       <Button size='large' variant='contained'>
         상담 신청하기
